+++ CHANGELOG.rst
@@ -8,6 +8,12 @@ Changelog
 
  .. note:: This version is not yet released and is under active development.
 
+* Added support for parsing PKCS12 files with friendly names for all
+  certificates with
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs12.load_pkcs12`,
+  which will return an object of type
+  :class:`~cryptography.hazmat.primitives.serialization.pkcs12.PKCS12KeyAndCertificates`.
+
 .. _v35-0-0:
 
 35.0.0 - 2021-09-29